adding exception handling for controllers
