+++ StudentAccounting/Controllers/RankController.cs
@@ -15,35 +15,77 @@ namespace StudentAccounting.Controllers
         [HttpGet("GetRank")]
         public ActionResult<IEnumerable<Rank>> Get()
         {
-            return Ok(_routingService.Get());
+            try
+            {
+                return Ok(_routingService.Get());
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
         [HttpGet("idRank/{id}", Name = "GetRankId")]
         public IActionResult Get(int id)
         {
-            return Ok(_routingService.Get(id));
+            try
+            {
+                return Ok(_routingService.Get(id));
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
         [HttpGet("nameRank/{name}", Name = "GetRankName")]
         public IActionResult Get(string name)
         {
-            return Ok(_routingService.Get(name));
+            try
+            {
+                return Ok(_routingService.Get(name));
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
         [HttpPost("CreateRank")]
         public IActionResult Create(Rank Rank)
         {
-            _routingService.Create(Rank);
-            return Ok();
+            try
+            {
+                _routingService.Create(Rank);
+                return Ok();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
         [HttpPut("UpdateRank")]
         public IActionResult Update(Rank Rank)
         {
-            _routingService.Edit(Rank);
-            return Ok();
+            try
+            {
+                _routingService.Edit(Rank);
+                return Ok();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
         [HttpDelete("DeleteRank")]
         public IActionResult Delete(int id)
         {
-             _routingService.Delete(id);
-            return Ok();
+            try
+            {
+                _routingService.Delete(id);
+                return Ok();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
     }
 }
